Create New File DialogFlow.js that will contain all the thing related to DialogFlow
